Working on relations
Many to many relation is now available. We can create that relation between tables and sql code is generated
correctly. One to many relation is supported in sql code. Program also checks if table names are unique
and if column names in every table are unique. If they are not unique program displays short notification.
to correct that mistake.